cookie provider tests in correct dir
